Extract player detection cone into its own script
