+++ Assets/Scripts/Enemy/Watcher.cs
@@ -12,16 +12,7 @@ public class Watcher : MonoBehaviour {
 	private static readonly int endFixLampHash = Animator.StringToHash("endFixLamp");
 	private static readonly int turnHash = Animator.StringToHash("Turn");
 
-	[SerializeField] private Vector2 eyeOffset = Vector2.zero;
-	[SerializeField] private float visionConeAngle = 30f;
-	[SerializeField] private float visionDistance = 4f;
-	[SerializeField, Range(1, 15)]
-	private int visionRayCount = 3;
-	[SerializeField] private LayerMask raycastLayers = -1;
-	[SerializeField] private bool ignoreTriggers = true;
-
-	[Space]
-	[SerializeField] private Transform lantern = null;
+	[SerializeField] private PlayerDetectionCone lantern;
 	[SerializeField] private Collider2D wallCollider = null;
 	[SerializeField] private Collider2D groundCollider = null;
 	[SerializeField] private LayerMask wallLayer = 1 << 8;
@@ -115,11 +106,10 @@ public class Watcher : MonoBehaviour {
 	}
 
 	private void Update() {
-		lantern.localPosition = eyeOffset;
 		facing = Mathf.Sign(transform.localScale.x);
 
 		playerPosition = playerTransform.position;
-		playerDirection = (playerPosition - lantern.position).normalized;
+		playerDirection = (playerPosition - lantern.transform.position).normalized;
 
 		bool isWalking = false;
 		if (doMovement && state != State.Recharging && state != State.Turning) {
@@ -144,7 +134,7 @@ public class Watcher : MonoBehaviour {
 			if (facing < 0) lanternAngle = 180f - lanternAngle;
 		}
 
-		lantern.localRotation = Quaternion.AngleAxis(lanternAngle, Vector3.forward);
+		lantern.Rotation = lanternAngle;
 		if (state != State.Recharging)
 			UpdatePlayerVisibility();
 
@@ -207,7 +197,7 @@ public class Watcher : MonoBehaviour {
 	}
 
 	private void UpdatePlayerVisibility() {
-		if (isPlayerVisible == (isPlayerVisible = CheckPlayerVisible())) return;
+		if (isPlayerVisible == (isPlayerVisible = state != State.Turning && lantern.IsPlayerVisible)) return;
 		if (isPlayerVisible) {
 			if (!isPlayerConsideredVisible) {
 				state = State.Following;
@@ -222,52 +212,6 @@ public class Watcher : MonoBehaviour {
 			if (looseVisibilityTimer.Elapsed)
 				looseVisibilityTimer.Reset(visibilityPadTime);
 		}
-	}
-
-	private bool CheckPlayerVisible() {
-		if (state == State.Turning) return false;
-
-		Vector2 eyePosition = transform.TransformPoint(eyeOffset);
-		Vector2 forward = lantern.right * facing;
-		float angleStep = visionConeAngle / (visionRayCount - 1);
-
-		for (int i = 0; i < visionRayCount; i++) {
-			float angle = -visionConeAngle / 2f + angleStep * i;
-			Vector2 direction = MathX.Rotate(forward, angle * Mathf.Deg2Rad);
-
-			if (DoSingleRaycast(eyePosition, direction)) {
-			#if UNITY_EDITOR
-				if (!Application.isPlaying) continue;
-			#endif
-
-				return true;
-			}
-		}
-
-		return false;
-	}
-
-	private bool DoSingleRaycast(Vector2 origin, Vector2 direction) {
-		RaycastHit2D hit = Physics2D.Raycast(origin, direction, visionDistance, raycastLayers);
-
-	#if UNITY_EDITOR
-		float distance = hit.collider != null ? hit.distance : visionDistance;
-		Debug.DrawRay(origin, direction * distance, Color.red);
-
-		if (hit.collider != null)
-			Debug.DrawLine(hit.point, hit.point + hit.normal * 0.1f, Color.yellow);
-	#endif
-
-		if (hit.rigidbody != null && hit.rigidbody.CompareTag("Player")) {
-			if (ignoreTriggers && hit.collider.isTrigger)
-				return false;
-
-			Transformation transformation = hit.rigidbody.GetComponent<Transformation>();
-			if (transformation == null || transformation.State != TransformationState.Human)
-				return true;
-		}
-
-		return false;
 	}
 
 	private void OnDetected() {
@@ -336,19 +280,7 @@ public class Watcher : MonoBehaviour {
 
 
 #if UNITY_EDITOR
-	private void OnDrawGizmos() {
-		if (!enabled || Application.isPlaying) return;
-		Vector3 eyePosition = transform.TransformPoint(eyeOffset);
-
-		Gizmos.color = Color.green;
-		Gizmos.DrawWireSphere(eyePosition, 0.2f);
-
-		CheckPlayerVisible();
-	}
-
 	private void OnValidate() {
-		lantern.localPosition = eyeOffset;
-
 		if (Application.isPlaying) {
 			Animator anim = GetComponent<Animator>();
 			if (anim.isInitialized)
